refactor: move error to cmd internal packages
