+++ cmd/oras/discover.go
@@ -30,6 +30,7 @@ import (
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	artifactspec "github.com/oras-project/artifacts-spec/specs-go/v1"
 	"github.com/spf13/cobra"
+	"oras.land/oras/cmd/oras/internal/errors"
 )
 
 type discoverOptions struct {
@@ -79,7 +80,7 @@ func runDiscover(opts discoverOptions) error {
 		return err
 	}
 	if repo.Reference.Reference == "" {
-		return newErrInvalidReference(repo.Reference)
+		return errors.NewErrInvalidReference(repo.Reference)
 	}
 
 	// discover artifacts
